refactor(xray-lite): move Context to context
Moves `Context` and `SubsegmentContext` to a new submodule `context`.
