+++ xray-lite/src/lib.rs
@@ -107,6 +107,7 @@
 //! ```
 
 mod client;
+mod context;
 mod epoch;
 mod error;
 mod header;
@@ -120,6 +121,7 @@ mod trace_id;
 
 pub use crate::{
     client::Client,
+    context::{Context, SubsegmentContext},
     epoch::Seconds,
     error::{Error, Result},
     header::Header,
@@ -129,66 +131,3 @@ pub use crate::{
     session::SubsegmentSession,
     trace_id::TraceId,
 };
-
-/// Context.
-pub trait Context {
-    /// Enters in a new subsegment.
-    ///
-    /// [`SubsegmentSession`] records the end of the subsegment when it is
-    /// dropped.
-    fn enter_subsegment<T>(&self, namespace: T) -> SubsegmentSession<T>
-    where
-        T: Namespace + Send + Sync;
-}
-
-/// Context as a subsegment of an existing segment.
-#[derive(Debug)]
-pub struct SubsegmentContext {
-    client: Client,
-    header: Header,
-    name_prefix: String,
-}
-
-impl SubsegmentContext {
-    /// Creates a new context from the Lambda environment variable.
-    ///
-    /// The following environment variable must be set:
-    /// - `_X_AMZN_TRACE_ID`: AWS X-Ray trace ID
-    ///
-    /// Please refer to the [AWS documentation](https://docs.aws.amazon.com/lambda/latest/dg/configuration-envvars.html#configuration-envvars-runtime)
-    /// for more details.
-    pub fn from_lambda_env(client: Client) -> Result<Self> {
-        let header = lambda::header()?;
-        Ok(Self {
-            client,
-            header,
-            name_prefix: "".to_string(),
-        })
-    }
-
-    /// Updates the context with a given name prefix.
-    ///
-    /// The name prefix is prepended to the name of every custom subsegment.
-    /// Only subsegments associated with [`CustomNamespace`] are affected.
-    pub fn with_name_prefix(self, prefix: impl Into<String>) -> Self {
-        Self {
-            client: self.client,
-            header: self.header,
-            name_prefix: prefix.into(),
-        }
-    }
-}
-
-impl Context for SubsegmentContext {
-    fn enter_subsegment<T>(&self, namespace: T) -> SubsegmentSession<T>
-    where
-        T: Namespace + Send + Sync,
-    {
-        SubsegmentSession::new(
-            self.client.clone(),
-            &self.header,
-            namespace,
-            &self.name_prefix,
-        )
-    }
-}
